// apps/sonarr-sync/src/generated/types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * A signed eight-byte integer. The upper big integer values are greater than the
   * max value for a JavaScript number. Therefore all big integers will be output as
   * strings and not numbers.
   */
  BigInt: any;
  /** A location in a connection that can be used for resuming pagination. */
  Cursor: any;
  /**
   * A point in time as described by the [ISO
   * 8601](https://en.wikipedia.org/wiki/ISO_8601) standard. May or may not include a timezone.
   */
  Datetime: string;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { [key: string]: any };
};

/** A connection to a list of `AnimeMetadatum` values. */
export type AnimeMetadataConnection = {
  __typename?: 'AnimeMetadataConnection';
  /** A list of edges which contains the `AnimeMetadatum` and cursor to aid in pagination. */
  edges: Array<AnimeMetadataEdge>;
  /** A list of `AnimeMetadatum` objects. */
  nodes: Array<Maybe<AnimeMetadatum>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The count of *all* `AnimeMetadatum` you could get from the connection. */
  totalCount: Scalars['Int'];
};
// ...
export type CreateSonarrSeryPayload = {
  __typename?: 'CreateSonarrSeryPayload';
  /**
   * The exact same `clientMutationId` that was provided in the mutation input,
   * unchanged and unused. May be used by a client to track mutations.
   */
  clientMutationId?: Maybe<Scalars['String']>;
  /** Our root query field type. Allows us to run any query from our mutation payload. */
  query?: Maybe<Query>;
  /** The `SonarrSery` that was created by this mutation. */
  sonarrSery?: Maybe<SonarrSery>;
  /** An edge for our `SonarrSery`. May be used by Relay 1. */
  sonarrSeryEdge?: Maybe<SonarrSeriesEdge>;
};
// ...
export type DownloadConfigsConnection = {
  __typename?: 'DownloadConfigsConnection';
  /** A list of edges which contains the `DownloadConfig` and cursor to aid in pagination. */
  edges: Array<DownloadConfigsEdge>;
  /** A list of `DownloadConfig` objects. */
  nodes: Array<Maybe<DownloadConfig>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The count of *all* `DownloadConfig` you could get from the connection. */
  totalCount: Scalars['Int'];
};
// ...
export type Node = {
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID'];
};
// ...
export type Query = Node & {
  __typename?: 'Query';
  /** Reads and enables pagination through a set of `AnimeMetadatum`. */
  allAnimeMetadata?: Maybe<AnimeMetadataConnection>;
  /** Reads and enables pagination through a set of `DownloadConfig`. */
  allDownloadConfigs?: Maybe<DownloadConfigsConnection>;
  /** Reads and enables pagination through a set of `RssItem`. */
  allRssItems?: Maybe<RssItemsConnection>;
  /** Reads and enables pagination through a set of `SonarrSery`. */
  allSonarrSeries?: Maybe<SonarrSeriesConnection>;
  /** Reads a single `AnimeMetadatum` using its globally unique `ID`. */
  animeMetadatum?: Maybe<AnimeMetadatum>;
  animeMetadatumById?: Maybe<AnimeMetadatum>;
  /** Reads a single `DownloadConfig` using its globally unique `ID`. */
  downloadConfig?: Maybe<DownloadConfig>;
  downloadConfigById?: Maybe<DownloadConfig>;
  /** Fetches an object given its globally unique `ID`. */
  node?: Maybe<Node>;
  /** The root query type must be a `Node` to work well with Relay 1 mutations. This just resolves to `query`. */
  nodeId: Scalars['ID'];
  /**
   * Exposes the root query type nested one level down. This is helpful for Relay 1
   * which can only query top level fields if they are in a particular form.
   */
  query: Query;
  /** Reads a single `RssItem` using its globally unique `ID`. */
  rssItem?: Maybe<RssItem>;
  rssItemByHash?: Maybe<RssItem>;
  rssItemById?: Maybe<RssItem>;
  /** Reads a single `SonarrSery` using its globally unique `ID`. */
  sonarrSery?: Maybe<SonarrSery>;
  sonarrSeryById?: Maybe<SonarrSery>;
  sonarrSeryBySonarrId?: Maybe<SonarrSery>;
  sonarrSeryBySonarrSlug?: Maybe<SonarrSery>;
};
// ...
export type RssItem = Node & {
  __typename?: 'RssItem';
  hash: Scalars['String'];
  id: Scalars['Int'];
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID'];
  publishDate: Scalars['Datetime'];
  size: Scalars['BigInt'];
  title: Scalars['String'];
  torrentLink: Scalars['String'];
};
// ...
export type RssItemsConnection = {
  __typename?: 'RssItemsConnection';
  /** A list of edges which contains the `RssItem` and cursor to aid in pagination. */
  edges: Array<RssItemsEdge>;
  /** A list of `RssItem` objects. */
  nodes: Array<Maybe<RssItem>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The count of *all* `RssItem` you could get from the connection. */
  totalCount: Scalars['Int'];
};
// ...
export type SonarrSeriesConnection = {
  __typename?: 'SonarrSeriesConnection';
  /** A list of edges which contains the `SonarrSery` and cursor to aid in pagination. */
  edges: Array<SonarrSeriesEdge>;
  /** A list of `SonarrSery` objects. */
  nodes: Array<Maybe<SonarrSery>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** The count of *all* `SonarrSery` you could get from the connection. */
  totalCount: Scalars['Int'];
};

/** A `SonarrSery` edge in the connection. */
export type SonarrSeriesEdge = {
  __typename?: 'SonarrSeriesEdge';
  /** A cursor for use in pagination. */
  cursor?: Maybe<Scalars['Cursor']>;
  /** The `SonarrSery` at the end of the edge. */
  node?: Maybe<SonarrSery>;
};
// ...
export type SonarrSery = Node & {
  __typename?: 'SonarrSery';
  /** Reads and enables pagination through a set of `AnimeMetadatum`. */
  animeMetadataBySonarrSeries: AnimeMetadataConnection;
  id: Scalars['Int'];
  isArchived: Scalars['Boolean'];
  /** A globally unique identifier. Can be used in various places throughout the system to identify this single value. */
  nodeId: Scalars['ID'];
  sonarrId: Scalars['Int'];
  sonarrName: Scalars['String'];
  sonarrSlug: Scalars['String'];
  tvdbid: Scalars['Int'];
};
// ...
export type SonarrSeryInput = {
  id?: InputMaybe<Scalars['Int']>;
  isArchived?: InputMaybe<Scalars['Boolean']>;
  sonarrId: Scalars['Int'];
  sonarrName: Scalars['String'];
  sonarrSlug: Scalars['String'];
  tvdbid: Scalars['Int'];
};

/** Represents an update to a `SonarrSery`. Fields that are set will be updated. */
export type SonarrSeryPatch = {
  id?: InputMaybe<Scalars['Int']>;
  isArchived?: InputMaybe<Scalars['Boolean']>;
  sonarrId?: InputMaybe<Scalars['Int']>;
  sonarrName?: InputMaybe<Scalars['String']>;
  sonarrSlug?: InputMaybe<Scalars['String']>;
  tvdbid?: InputMaybe<Scalars['Int']>;
};
// ...
export type CreateSonarrSeriesMutationVariables = Exact<{
  sonarrSery: SonarrSeryInput;
}>;


export type CreateSonarrSeriesMutation = { __typename?: 'Mutation', createSonarrSery?: { __typename: 'CreateSonarrSeryPayload' } | null | undefined };

export type GetSonarrSeriesByIdListQueryVariables = Exact<{
  in?: InputMaybe<Array<Scalars['Int']> | Scalars['Int']>;
}>;


export type GetSonarrSeriesByIdListQuery = { __typename?: 'Query', allSonarrSeries?: { __typename?: 'SonarrSeriesConnection', edges: Array<{ __typename?: 'SonarrSeriesEdge', node?: { __typename?: 'SonarrSery', sonarrId: number, sonarrName: string, sonarrSlug: string, tvdbid: number, id: number } | null | undefined }> } | null | undefined };

export type UpdateSonarrSeriesMutationVariables = Exact<{
  id: Scalars['Int'];
  sonarrSeryPatch: SonarrSeryPatch;
}>;


export type UpdateSonarrSeriesMutation = { __typename?: 'Mutation', updateSonarrSeryById?: { __typename: 'UpdateSonarrSeryPayload' } | null | undefined };


export const CreateSonarrSeriesDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"mutation","name":{"kind":"Name","value":"CreateSonarrSeries"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"sonarrSery"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"SonarrSeryInput"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"createSonarrSery"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"input"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"sonarrSery"},"value":{"kind":"Variable","name":{"kind":"Name","value":"sonarrSery"}}}]}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}}]}}]}}]} as unknown as DocumentNode<CreateSonarrSeriesMutation, CreateSonarrSeriesMutationVariables>;
export const GetSonarrSeriesByIdListDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"query","name":{"kind":"Name","value":"GetSonarrSeriesByIdList"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"in"}},"type":{"kind":"ListType","type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"allSonarrSeries"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"filter"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"sonarrId"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"in"},"value":{"kind":"Variable","name":{"kind":"Name","value":"in"}}}]}}]}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"edges"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"node"},"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"sonarrId"}},{"kind":"Field","name":{"kind":"Name","value":"sonarrName"}},{"kind":"Field","name":{"kind":"Name","value":"sonarrSlug"}},{"kind":"Field","name":{"kind":"Name","value":"tvdbid"}},{"kind":"Field","name":{"kind":"Name","value":"id"}}]}}]}}]}}]}}]} as unknown as DocumentNode<GetSonarrSeriesByIdListQuery, GetSonarrSeriesByIdListQueryVariables>;
export const UpdateSonarrSeriesDocument = {"kind":"Document","definitions":[{"kind":"OperationDefinition","operation":"mutation","name":{"kind":"Name","value":"UpdateSonarrSeries"},"variableDefinitions":[{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"id"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"Int"}}}},{"kind":"VariableDefinition","variable":{"kind":"Variable","name":{"kind":"Name","value":"sonarrSeryPatch"}},"type":{"kind":"NonNullType","type":{"kind":"NamedType","name":{"kind":"Name","value":"SonarrSeryPatch"}}}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"updateSonarrSeryById"},"arguments":[{"kind":"Argument","name":{"kind":"Name","value":"input"},"value":{"kind":"ObjectValue","fields":[{"kind":"ObjectField","name":{"kind":"Name","value":"sonarrSeryPatch"},"value":{"kind":"Variable","name":{"kind":"Name","value":"sonarrSeryPatch"}}},{"kind":"ObjectField","name":{"kind":"Name","value":"id"},"value":{"kind":"Variable","name":{"kind":"Name","value":"id"}}}]}}],"selectionSet":{"kind":"SelectionSet","selections":[{"kind":"Field","name":{"kind":"Name","value":"__typename"}}]}}]}}]} as unknown as DocumentNode<UpdateSonarrSeriesMutation, UpdateSonarrSeriesMutationVariables>;